+++ src/middleware/tenant.middleware.ts
@@ -7,42 +7,7 @@ import { Request, Response, NextFunction } from 'express';
 import { PinoLogger } from 'nestjs-pino';
 import { ConfigService } from '@nestjs/config';
 import { TenantService } from '../user-management/tenant.service';
-
-async function createSessionToken( // this function is for testing only, should be removed in production
-  sessionId: string,
-  clerkSecretKey: string,
-  logger: PinoLogger,
-) {
-  logger.debug({ sessionId, clerkSecretKey }, 'Creating session token');
-
-  try {
-    const response = await fetch(
-      'https://api.clerk.com/v1/sessions/' + sessionId + '/tokens',
-      {
-        method: 'POST',
-        headers: {
-          Authorization: `Bearer ${clerkSecretKey}`,
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({
-          expires_in_seconds: 9600,
-        }),
-      },
-    );
-
-    if (!response.ok) {
-      const errorDetails = await response.json();
-      throw new Error(`Error ${response.status}: ${errorDetails.message}`);
-    }
-
-    const data = await response.json();
-    logger.debug(`Session Token: ${data.jwt}`);
-    return data.jwt;
-  } catch (error) {
-    logger.error('Failed to create session token:', error.message);
-    throw error;
-  }
-}
+import { createSessionToken } from '../utils/createSessionToken';
 
 @Injectable()
 export class TenantMiddleware implements NestMiddleware {
@@ -57,12 +22,9 @@ export class TenantMiddleware implements NestMiddleware {
     const clerkSecretKey = this.configService.get('clerk.secretKey');
 
     // testing: renew session by 1 hr
-    const customJWT = await createSessionToken(
-      sessionId,
-      clerkSecretKey,
-      this.logger,
-    );
-    req.headers['authorization'] = `Bearer ${customJWT}`;
+    if (process.env.NODE_ENV === 'development')
+      await createSessionToken(sessionId, clerkSecretKey, this.logger);
+    // req.headers['authorization'] = `Bearer ${customJWT}`;
     const tenantId =
       (req.headers['X-TENANT-ID'] as string) ||
       (req.headers['x-tenant-id'] as string);
