+++ src/platforms/linkedin/repositories/linkedin.repository.spec.ts
@@ -181,25 +181,16 @@ describe('LinkedInRepository', () => {
         text: 'Test post',
         organizationId: 'org-id',
       } as unknown as LinkedInPost;
-      const createdPost = {
-        id: 'temp-id',
-        ...postData,
-      } as unknown as LinkedInPost;
-      const savedPost = {
-        id: 'new-post-id',
-        ...postData,
-      } as unknown as LinkedInPost;
-
-      jest.spyOn(postRepo, 'create').mockReturnValue(createdPost);
-      jest.spyOn(postRepo, 'save').mockResolvedValue(savedPost);
+      postRepo.create.mockReturnValue(postData);
+      postRepo.save.mockResolvedValue({ id: 'new-post-id', ...postData });
 
       // Act
       const result = await repository.createPost(postData);
 
       // Assert
       expect(postRepo.create).toHaveBeenCalledWith(postData);
-      expect(postRepo.save).toHaveBeenCalledWith(createdPost);
-      expect(result).toEqual(savedPost);
+      expect(postRepo.save).toHaveBeenCalled();
+      expect(result).toEqual({ id: 'new-post-id', ...postData });
     });
   });
